+++ changelog.md
@@ -1,2 +1,7 @@
 ## Improved:
 - differentiate between version branch `stable` and `preview`
+- function get-online-version() store seen versions in json-file
+
+## New:
+- function update-server()
+- function update-all-server()